Refactoring: breaking down in fields and utils
