+++ tox.ini
@@ -26,7 +26,18 @@ addopts =
 pep8ignore = * ALL
 
 [testenv]
-passenv = TRAVIS TRAVIS_JOB_ID TRAVIS_BRANCH PYTHONDONTWRITEBYTECODE
+passenv =
+    - TRAVIS
+    - TRAVIS_JOB_ID
+    - TRAVIS_BRANCH
+    - PYTHONDONTWRITEBYTECODE
+    - PGDATABASE
+    - PGHOSTADDR
+    - PGHOST
+    - PGPORT
+    - PGUSER
+    - PGPASSWORD
+    - PGPASS
 
 whitelist_externals =
                     /usr/local/bin/psql
